+++ docs/internal/investigations/runtime-governor-poc/component/go.mod
@@ -1,4 +1,4 @@
-module github.com/kyma-project/control-plane/docs/internal/investigations/runtime-director-poc/component
+module github.com/kyma-project/control-plane/docs/internal/investigations/runtime-governor-poc/component
 
 go 1.14
