+++ docs/internal/investigations/runtime-governor-poc/component/pkg/model/model.go
